refactor: move video schema definitions to dedicated schema directory
