+++ src/apps/io/videos/index.ts
@@ -1,13 +1,13 @@
 import express, { Router } from 'express';
 import { validateRequest } from 'src/utils/validator';
 import { fixDurationHandler } from './routes/fix-duration';
-import { FixDurationHandlerRequest, FixDurationHandlerSchema } from './routes/fix-duration/schema';
 import { fixThumbnailHandler } from './routes/fix-thumbnail';
-import { FixThumbnailHandlerRequest, FixThumbnailHandlerSchema } from './routes/fix-thumbnail/schema';
 import { importPlatformHandler } from './routes/import-platform';
 import { streamHLSHandler } from './routes/stream-hls';
 import { StreamHandlerRequest, streamHandlerSchema } from 'src/schema/videos/stream-hls';
 import { ImportHandlerRequest, importHandlerSchema } from 'src/schema/videos/import-platform';
+import { FixDurationHandlerRequest, fixDurationHandlerSchema } from 'src/schema/videos/fix-duration';
+import { FixThumbnailHandlerRequest, fixThumbnailHandlerSchema } from 'src/schema/videos/fix-thumbnail';
 
 const videosRouter: Router = express.Router();
 
@@ -19,12 +19,12 @@ videosRouter.post(
 );
 videosRouter.post(
   '/fix-duration',
-  validateRequest<FixDurationHandlerRequest>(FixDurationHandlerSchema),
+  validateRequest<FixDurationHandlerRequest>(fixDurationHandlerSchema),
   fixDurationHandler
 );
 videosRouter.post(
   '/fix-thumbnail',
-  validateRequest<FixThumbnailHandlerRequest>(FixThumbnailHandlerSchema),
+  validateRequest<FixThumbnailHandlerRequest>(fixThumbnailHandlerSchema),
   fixThumbnailHandler
 );
