+++ src/schema/videos/fix-duration/index.ts
@@ -2,12 +2,12 @@ import { z } from 'zod';
 import { taskHandlerHeaderSchema } from 'src/utils/cloud-task/schema';
 import { videoDataSchema } from 'src/schema/videos/convert';
 
-const FixThumbnailHandlerSchema = z.object({
+const fixDurationHandlerSchema = z.object({
   body: z.object({
     id: videoDataSchema.shape.id,
   }),
   headers: taskHandlerHeaderSchema.passthrough(),
 });
 
-export type FixThumbnailHandlerRequest = z.infer<typeof FixThumbnailHandlerSchema>;
-export { FixThumbnailHandlerSchema };
+export type FixDurationHandlerRequest = z.infer<typeof fixDurationHandlerSchema>;
+export { fixDurationHandlerSchema };
